cuckoo: address review + move CuckooMap to proto, split errors, siphash + seed on wire, refresh docs
