+++ Cargo.toml
@@ -59,6 +59,7 @@ rand = "0.8"
 rayon = "1.11.0"
 serde = "1.0.145"
 serde_json = "1.0.86"
+siphasher = "1"
 smallvec = "1.15.1"
 thiserror = "2.0.16"
 tokio = "1.47.1"
