+++ yellowstone-grpc-proto/Cargo.toml
@@ -14,6 +14,7 @@ publish = true
 bytes = { workspace = true, optional = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
+siphasher ={ workspace = true}
 tonic = { workspace = true, optional = true }
 tonic-prost = { workspace = true, optional = true }
 thiserror = { workspace = true }
